Sebastien's threaded tests for marvin
While this sends out workers for each test fixture there seem to be
problems with urllib2 and threading. urllib2 being unsafe the threads
behaviour is not deterministic. Tests timeout early or are unable to
contact the CloudStack end point. Retaining the patch to test with a
more thread safe library