debian: Re-do packaging
The packaging way for Debian has been re-done in this commit.

While this is a commit in a feature branch it is not final yet.

The idea is that packages are build by running: "dpkg-buildpackage" which calls debian/rules

That calls maven again which calls the artifacts.

Configuration for agent, client and server have been moved into /conf which is a seperate maven project.

This only builds configuration artifacts based on build/replace.properties which is filled by debian/rules

The same should be done for files in agent/bindir, client/bindir, etc. These tools can be a own maven target,
they don't have a relation with the Java code.

All dependencies for the Agent, Client and Server are packed into the WAR/JAR files. This makes the packages bigger,
but it makes sure they run on all platforms without having to rely on system dependencies.

cloud-client(-ui) has been merged into cloud-server and packages like cloud-agent-libs, cloud-deps and cloud-agent-deps
have been deprecated due to the build-in dependencies.

This reduces the number of packages, but also makes integration nicer.

A lot of work is still to be done:
* Tools like cloud-setup-databases, etc
* postinst / preinst files shouldn't touch anything (like /etc/limits.conf)

This is just a PoC to show what can be done