kvm: During migrate change the VNC listen address
Commit 7240204a507cce8143c248e6aa635da6dad60ed0 re-added the functionality that
the VNC on KVM hypervisors would listen on the private IP address of the hypervisor.

This broke migrations since Qemu on the target hypervisor would try to bind to the
IP address of the old hypervisor.

The migrate method from libvirt supports passing down a different XML for running
the instance of the target hypervisor.

A modification in libvirt-java was required for this. Without this modification
in libvirt-java this code won't compile.

The assumption is that libvirt-java 0.50.0 will have the required fixes.